Move the 'FileNode' class into a separate file
